ai-search/pyproject.toml
[project]
name = "ai-search"
version = "1.0.2"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deep-translator>=1.11.4",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langchain-google-genai>=2.0.0",
    "langchain-tavily>=0.2.11",
    "openai>=1.107.3",
    "google-generativeai>=0.8.0",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.2",
    "streamlit>=1.49.1",
    "tavily-python>=0.7.12",
]
